Fix armv7 build and cross-compile every published platform in CI
Statfs_t.Bsize is int64 on 64-bit platforms and int32 on 32-bit ones,
so the free-space calculation compiled on amd64 and arm64 and failed on
armv7. Both operands now converted explicitly.

The check job cross-compiled for Windows and macOS but not for any
platform we actually publish, so this only surfaced in the slow
multi-arch image build. It now builds all five targets in the fast job,
where a type error costs seconds instead of two minutes of QEMU.

+++ internal/plex/diskfree_linux.go
@@ -11,5 +11,8 @@ func freeBytes(path string) int64 {
 	if err := syscall.Statfs(path, &st); err != nil {
 		return 0
 	}
-	return int64(st.Bavail) * st.Bsize
+	// Both operands need an explicit conversion: Bsize is int64 on 64-bit
+	// platforms but int32 on 32-bit ones (armv7), and Bavail is unsigned. A
+	// bare multiplication compiles on amd64 and fails on arm.
+	return int64(st.Bavail) * int64(st.Bsize)
 }
